CAP-230 resolve request change
